Fix inappropriate default for certs directory
I get a permission denied error when trying to access a secure registry:

Error: open /etc/docker/certs.d/foobar: permission denied

This is because I am not permitted to access this folder. Since it's
configuration specifically for the docker engine, I think it's reasonable
that we take responsibility for it's value.